refactor to test_util file
