+++ src/cfg/tests.rs
@@ -1,3 +1,4 @@
+use crate::test_util::*;
 use crate::{
     cfg::{function_to_cfg, program_to_cfg, to_structured::cfg_to_structured, BlockName},
     EggCCError,
@@ -13,55 +14,6 @@ fn parse_from_string(input: &str) -> Program {
     let json_str = String::from_utf8(buf).unwrap();
     load_program_from_read(json_str.as_bytes())
 }
-
-macro_rules! to_block {
-    (ENTRY) => {
-        BlockName::Entry
-    };
-    (EXIT) => {
-        BlockName::Exit
-    };
-    ($name:expr) => {
-        BlockName::Named($name.into())
-    };
-}
-pub(crate) use to_block;
-
-macro_rules! cfg_test_equiv {
-    // for the case of a single-node cfg
-    ($cfg:expr, []) => {
-        assert_eq!($cfg.graph.node_count(), 1);
-        assert_eq!($cfg.graph.edge_count(), 0);
-        assert_eq!($cfg.entry, $cfg.exit);
-    };
-    ($cfg:expr,  [ $($src:tt = ($($edge:tt)*)=> $dst:tt,)* ]) => {
-        use $crate::cfg::BranchOp;
-        let mut mentioned = std::collections::HashSet::new();
-            let mut block = std::collections::HashMap::new();
-            $(
-                mentioned.insert(to_block!($src));
-                mentioned.insert(to_block!($dst));
-            )*
-            let cfg = $cfg;
-            for i in cfg.graph.node_indices() {
-                let node = cfg.graph.node_weight(i).unwrap();
-                assert!(mentioned.contains(&node.name), "description does not mention block {:?}", node.name);
-                block.insert(node.name.clone(), i);
-            }
-            $({
-                let src_name = to_block!($src);
-                let dst_name = to_block!($dst);
-                let src = *block.get(&src_name).unwrap_or_else(|| panic!("missing block {:?}", src_name));
-                let dst = *block.get(&dst_name).unwrap_or_else(|| panic!("missing block {:?}", dst_name));
-                let has_edge = cfg.graph.edges_connecting(src, dst).any(|edge| {
-                    edge.weight().op == BranchOp::$($edge)*
-                });
-                assert!(has_edge, "missing edge from {src_name:?} to {dst_name:?}");
-            })*
-    };
-}
-
-pub(crate) use cfg_test_equiv;
 
 // Test that a CFG is wired up correctly.
 macro_rules! cfg_test_function_to_cfg {
